Fix adding duplicate overlays when reverting buffer and in general
When a buffer is reverted (or when the major mode is reloaded), buffer
local variables are cleared, but overlays aren't. This causes
previously added overlays to linger around.

This commit marks all overlays added by scopeline with a property, and
makes it so that if we attempt to add a duplicate overlay, the
previously added ones would be deleted instead.

This commit does not make delete-all-overlays also eagerly look
through every overlay in the buffer and try to clean them up, because
that seems slow and unnecessary.